Fix and order types
