+++ front/app/services/projectFolderImages.ts
@@ -1,21 +1,42 @@
-import { ImageSizes } from 'typings';
 import { API_PATH } from 'containers/App/constants';
 import streams, { IStreamParams } from 'utils/streams';
+import { TProjectFolderCardSize } from 'components/ProjectAndFolderCards/components/ProjectFolderCard';
+import { CARD_IMAGE_ASPECT_RATIO as PROJECT_CARD_IMAGE_ASPECT_RATIO } from 'services/projectImages';
 
 const apiEndpoint = `${API_PATH}/project_folders`;
+
+export const CARD_IMAGE_ASPECT_RATIO = PROJECT_CARD_IMAGE_ASPECT_RATIO;
+
+export type ProjectFolderImageSizes = {
+  small: string | null;
+  large: string | null;
+};
+
+export const getCardImageUrl = (
+  imageVersions: ProjectFolderImageSizes,
+  isPhone: boolean,
+  size?: TProjectFolderCardSize
+) => {
+  if (isPhone || size !== 'small') {
+    // image size is approximately the same for both medium and large desktop card sizes
+    return imageVersions.large;
+  } else {
+    return imageVersions.small;
+  }
+};
 
 export interface IProjectFolderImageData {
   id: string;
   type: string;
   attributes: {
-    versions: ImageSizes;
+    versions: ProjectFolderImageSizes;
     ordering: string | null;
     created_at: string;
     updated_at: string;
   };
 }
 
-export interface IProjectFolderImage {
+interface IProjectFolderImage {
   data: IProjectFolderImageData;
 }
