+++ front/app/services/projectImages.ts
@@ -1,21 +1,44 @@
-import { ImageSizes } from 'typings';
 import { API_PATH } from 'containers/App/constants';
 import streams, { IStreamParams } from 'utils/streams';
+import { TProjectCardSize } from 'components/ProjectCard';
 
 const apiEndpoint = `${API_PATH}/projects`;
+
+type ProjectImageSizes = {
+  small: string | null;
+  large: string | null;
+};
+
+export const CARD_IMAGE_ASPECT_RATIO_WIDTH = 4;
+export const CARD_IMAGE_ASPECT_RATIO_HEIGHT = 3;
+export const CARD_IMAGE_ASPECT_RATIO =
+  CARD_IMAGE_ASPECT_RATIO_WIDTH / CARD_IMAGE_ASPECT_RATIO_HEIGHT;
+
+export const getCardImageUrl = (
+  imageVersions: ProjectImageSizes,
+  isPhone: boolean,
+  size?: TProjectCardSize
+) => {
+  if (isPhone || size !== 'small') {
+    // image size is approximately the same for both medium and large desktop card sizes
+    return imageVersions.large;
+  } else {
+    return imageVersions.small;
+  }
+};
 
 export interface IProjectImageData {
   id: string;
   type: string;
   attributes: {
-    versions: ImageSizes;
+    versions: ProjectImageSizes;
     ordering: string | null;
     created_at: string;
     updated_at: string;
   };
 }
 
-export interface IProjectImage {
+interface IProjectImage {
   data: IProjectImageData;
 }
